requirements/base.txt
Django==1.10
psycopg2==2.7.3.1
django-allauth==0.33.0
Pillow==4.3.0
django-stdimage==2.4.2
django-image-cropping==1.1.0
easy-thumbnails==2.4.2